% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nreps_classification.R
\name{calc_nreps_classification}
\alias{calc_nreps_classification}
\title{Determine sample sizes for a set of ML algorithms on a single data instance}
\usage{
calc_nreps_classification(
  instance,
  algorithms,
  se.max = 0.001,
  dif = "simple",
  comparisons = "all.vs.all",
  method = "param",
  nstart = 10,
  nmax = 500,
  seed = NULL,
  boot.R = 499,
  ncpus = 1,
  force.balanced = FALSE,
  load.folder = NA,
  save.folder = NA
)
}
\arguments{
\item{instance}{a list object containing the definitions of the problem
instance.
See Section `Instance` for details.}

\item{algorithms}{a list object containing the definitions of all algorithms.
See Section `Algorithms` for details.}

\item{se.max}{desired upper limit for the standard error of the estimated
difference between pairs of algorithms. See Section
`Pairwise Differences` for details.}

\item{dif}{type of difference to be used. Accepts "perc" (for percent
differences) or "simple" (for simple differences)}

\item{comparisons}{type of comparisons being performed. Accepts "all.vs.first"
(in which cases the first object in `algorithms` is considered to be
the reference algorithm) or "all.vs.all" (if there is no reference
and all pairwise comparisons are desired).}

\item{method}{method to use for estimating the standard errors. Accepts
"param" (for parametric) or "boot" (for bootstrap)}

\item{nstart}{initial number of algorithm runs for each algorithm.
See Section `Initial Number of Observations` for details.}

\item{nmax}{maximum **total** allowed number of runs to execute. Loaded
results (see `load.folder` below) do not count towards this
total.}

\item{seed}{seed for the random number generator}

\item{boot.R}{number of bootstrap resamples to use (if `method == "boot"`)}

\item{ncpus}{number of cores to use}

\item{force.balanced}{logical flag to force the use of balanced sampling for
the algorithms on each instance}

\item{load.folder}{name of folder to load results from. Use either "" or
"./" for the current working directory. Accepts relative paths.
Use `NA` for not saving. `calc_nreps()` will look for a .RDS file
with the same name}

\item{save.folder}{name of folder to save the results. Use either "" or
"./" for the current working directory. Accepts relative paths.
Use `NA` for not saving.}
}
\value{
a list object containing the following items:
\itemize{
   \item \code{instance} - alias for the problem instance considered
   \item \code{Xk} - list of observed performance values for all `algorithms`
   \item \code{Nk} - vector of sample sizes generated for each algorithm
   \item \code{Diffk} - data frame with point estimates, standard errors and
   other information for all algorithm pairs of interest
   \item \code{seed} - seed used for the PRNG
   \item \code{dif} - type of difference used
   \item \code{method} - method used ("param" / "boot")
   \item \code{comparisons} - type of pairings ("all.vs.all" / "all.vs.first")
}
}
\description{
Iteratively calculates the required sample sizes for K ML algorithms
on a given problem instance, so that the standard errors of the estimates of
the pairwise differences in performance is controlled at a predefined level.
}
\section{Instance}{

}

\section{Algorithms}{

}

\section{Initial Number of Observations}{

In the **general case** the initial number of observations per algorithm
(`nstart`) should be relatively high. For the parametric case
we recommend between 10 and 20 if outliers are not expected, or between 30
and 50 if that assumption cannot be made. For the bootstrap approach we
recommend using at least 20. However, if some distributional assumptions can
be made - particularly low skewness of the population of algorithm results on
the test instances), then `nstart` can in principle be as small as 5 (if the
output of the algorithms were known to be normal, it could be 1).

In general, higher sample sizes are the price to pay for abandoning
distributional assumptions. Use lower values of `nstart` with caution.
}

\section{Pairwise Differences}{

Parameter `dif` informs the type of difference in performance to be used
for the estimation (\eqn{\mu_a} and \eqn{\mu_b} represent the mean
performance of any two algorithms on the test instance, and \eqn{mu}
represents the grand mean of all algorithms given in `algorithms`):

- If `dif == "perc"` and `comparisons == "all.vs.first"`, the estimated quantity is
   \eqn{\phi_{1b} = (\mu_1 - \mu_b) / \mu_1 = 1 - (\mu_b / \mu_1)}.

- If `dif == "perc"` and `comparisons == "all.vs.all"`, the estimated quantity is
   \eqn{\phi_{ab} = (\mu_a - \mu_b) / \mu}.

- If `dif == "simple"` it estimates \eqn{\mu_a - \mu_b}.
}

\examples{
preprocess <- function(data)
{
  X <- data[,-c(1,2,4)]
  y <- data[,4]

  dat <- as.data.frame(X)
  dat$target <- y

  return(dat)
}


alphavirus <- preprocess(read.csv("Alphavirus.csv"))
instance <- list(alias = "Alphavirus", Data = alphavirus)
algorithms <- c("svm", "random_forest")
result <- calc_nreps_classification(instance, algorithms)
summary(result)
plot(result)
}
\references{
- F. Campelo
   CAISEr: Comparing Algorithms with Iterative Sample-size Estimation in R
- F. Campelo, F. Takahashi:
   Sample size estimation for power and accuracy in the experimental
   comparison of algorithms. Journal of Heuristics 25(2):305-338, 2019.
- P. Mathews.
   Sample size calculations: Practical methods for engineers and scientists.
   Mathews Malnar and Bailey, 2010.
-  A.C. Davison, D.V. Hinkley:
   Bootstrap methods and their application. Cambridge University Press (1997)
-  E.C. Fieller:
    Some problems in interval estimation. Journal of the Royal Statistical
    Society. Series B (Methodological) 16(2), 175–185 (1954)
- V. Franz:
   Ratios: A short guide to confidence limits and proper use (2007).
   https://arxiv.org/pdf/0710.2024v1.pdf
- D.C. Montgomery, C.G. Runger:
   Applied Statistics and Probability for Engineers, 6th ed. Wiley (2013)
}
\author{
Prajwel Venkatasubramaniyan (\email{prajwel16@gmail.com})
}
