% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmlaiser.R
\name{plot.CAISEr}
\alias{plot.CAISEr}
\title{plot.CMLAISEr}
\usage{
\method{plot}{CAISEr}(
  x,
  y = NULL,
  ...,
  latex = FALSE,
  reorder = FALSE,
  show.text = TRUE,
  digits = 3,
  layout = NULL
)
}
\arguments{
\item{x}{list object of class _CMLAISEr_.}

\item{y}{unused. Included for consistency with generic `plot` method.}

\item{...}{other parameters to be passed down to specific
plotting functions (currently unused)}

\item{latex}{logical: should labels be formatted for LaTeX? (useful for
later saving using library `TikzDevice`)}

\item{reorder}{logical: should the comparisons be reordered alphabetically?}

\item{show.text}{logical: should text be plotted?}

\item{digits}{how many significant digits should be used in text?}

\item{layout}{optional parameter to override the layout of the plots (see
`gridExtra::arrangeGrobs()` for details. The default layout is
`lay = rbind(c(1,1,1,1,1,1), c(1,1,1,1,1,1), c(2,2,2,3,3,3))`}
}
\value{
list containing (1) a list of of `ggplot2` objects generated, and
        (2) a list of data frames used for the creation of the plots.
}
\description{
S3 method for plotting _CMLAISEr_ objects output by [run_experiment()].
}
\section{References}{

- F. Campelo
   CAISEr: Comparing Algorithms with Iterative Sample-size Estimation in R
}

