% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_nreps.R
\name{summary.nreps}
\alias{summary.nreps}
\title{summary.reps}
\usage{
\method{summary}{nreps}(object, ...)
}
\arguments{
\item{object}{list object of class _nreps_
(generated by [adaptive_reps()])}

\item{...}{other parameters to be passed down to specific
summary functions (currently unused)}
}
\description{
S3 method for summarizing _nreps_ objects output by [adaptive_nreps()]).
}
\section{References}{

- F. Campelo
   CAISEr: Comparing Algorithms with Iterative Sample-size Estimation in R
}

